Pricing and Channel Strategy Analysis for Premium Product
This project involves pricing analysis and product strategy in a complex product pricing and channel strategy scenario. This case study dives deep into the effects of aggressive discounting and a fixed dealer pricing on margins, brand positioning, and dealer incentives.

In this study I have applied cost and contribution margin analysis, also explored the breakeven impacts, and have given my strategic recommendations that balances fixed pricing with flexible minimum advertised pricing (MAP). The goal for this study was to preserve the product's premium status and to enhance dealer collaboration and also to better the revenue performance.

With projects like this I am eager to leverage my knowledge in pricing, cost management and channel strategy in real-world consulting and pricing strategy roles.

#PricingStrategy #CostAnalysis #ProductStrategy #RevenueManagement